Move vote session to Features namespace
